launcher

+++ data/api/schema.php
@@ -2,7 +2,8 @@
 
 include_once $_SERVER["DOCUMENT_ROOT"] . "/db/db.php";
 
-query("DROP TABLE IF EXISTS `data`;");
+if (DEBUG)
+    query("DROP TABLE IF EXISTS `data`;");
 query("CREATE TABLE IF NOT EXISTS `data` (
   `data_id` int(11) NOT NULL AUTO_INCREMENT,
   `data_parent_id` int(11) DEFAULT NULL,
@@ -12,7 +13,8 @@ query("CREATE TABLE IF NOT EXISTS `data` (
   PRIMARY KEY (`data_id`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8 COLLATE=utf8_bin;");
 
-query("DROP TABLE IF EXISTS `history`;");
+if (DEBUG)
+    query("DROP TABLE IF EXISTS `history`;");
 query("CREATE TABLE IF NOT EXISTS `history` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `data_path` varchar(64) COLLATE utf8_bin NOT NULL,
@@ -21,7 +23,8 @@ query("CREATE TABLE IF NOT EXISTS `history` (
   PRIMARY KEY (`id`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8 COLLATE=utf8_bin;");
 
-query("DROP TABLE IF EXISTS `hashes`;");
+if (DEBUG)
+    query("DROP TABLE IF EXISTS `hashes`;");
 query("CREATE TABLE IF NOT EXISTS `hashes` (
   `hash` varchar(64) COLLATE utf8_bin NOT NULL,
   `path` varchar(255) COLLATE utf8_bin NOT NULL
